Move validation function back into lib/ansible/playbook/base.py
ci_complete
